requirements.txt
streamlit>=1.32.0
pandas>=2.2.1
numpy>=1.26.4
vectorbt>=0.28.0
duckdb>=0.10.1
pyarrow>=15.0.0
litellm>=1.34.19
langchain>=0.1.13
requests>=2.31.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.28
mcp>=1.0.0
